Usuarios: Fully protected API endpoints

+++ src/main/java/com/easytemplates/backend/controller/UsuarioController.java
@@ -1,10 +1,13 @@
 package com.easytemplates.backend.controller;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.easytemplates.backend.dto.Role;
 import com.easytemplates.backend.dto.Usuarios;
 import com.easytemplates.backend.service.UsuarioServiceImpl;
 
@@ -41,9 +45,17 @@ public class UsuarioController {
 		
 		Usuarios user = usuarioServiceImpl.usuarioById(id);
 		
-		return new ResponseEntity<Object>(user, HttpStatus.OK);
+		if (user != null)
+		{
+			return new ResponseEntity<Object>(user, HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<Object>("The specified user doesn\'t exist in the database, try registering it", HttpStatus.INTERNAL_SERVER_ERROR);
 	}
 
+	/**
+	 *  The user will have to login again (JWT Bearer Token is invalid after new user details)
+	 */
 	@PutMapping("/usuarios/{id}")
 	public ResponseEntity<Object> actualizarUsuario(@PathVariable(name = "id") Long id, @RequestBody UserDetailsRequestModel requestUserDetails) {
 		Usuarios user = usuarioServiceImpl.usuarioById(id);
@@ -53,8 +65,16 @@ public class UsuarioController {
 			String targetUserEmail = usuarioServiceImpl.usuarioById(id).getEmail();
 			String authedUserEmail = ((Usuarios) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getEmail();
 			
+			Set<Role> userRoles = ((Usuarios) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getRoles();
 			
-			if (targetUserEmail.equals(authedUserEmail))
+			HashSet<String> userRolesArray = new HashSet<String>(userRoles.size());
+			
+			// Map each role name to a Strings HashSet
+			for(Role role : userRoles) {
+				userRolesArray.add(role.toString());
+			}
+			
+			if (targetUserEmail.equals(authedUserEmail) || userRolesArray.contains("ROLE_ADMIN") == true)
 			{
 				String encoded = bCryptPasswordEncoder.encode(requestUserDetails.getPassword());
 				
@@ -75,9 +95,22 @@ public class UsuarioController {
 		return new ResponseEntity<Object>("The specified user doesn\'t exist in the database, try registering it", HttpStatus.INTERNAL_SERVER_ERROR);
 	}
 
+	/**
+	 *  The user will have to login again (JWT Bearer Token is invalid)
+	 */
+	@PreAuthorize("hasRole('ADMIN')")
 	@DeleteMapping("/usuarios/{id}")
-	public void eliminarUsuario(@PathVariable(name = "id") Long id) {
-		usuarioServiceImpl.deleteUsuario(id);
+	public ResponseEntity<Object> eliminarUsuario(@PathVariable(name = "id") Long id) {
+		Usuarios user = usuarioServiceImpl.usuarioById(id);
+		
+		if (user != null)
+		{
+			usuarioServiceImpl.deleteUsuario(id);
+			return new ResponseEntity<Object>("User deleted successfully", HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<Object>("The specified user doesn\'t exist in the database, won't delete", HttpStatus.INTERNAL_SERVER_ERROR);
+		
 	}
 	
 }
